+++ druid-shell/src/platform/x11/window.rs
@@ -522,22 +522,65 @@ impl Window {
         self.render(rect)
     }
 
+    fn mouse_button(&self, button: u8) -> MouseButton {
+        match button {
+            1 => MouseButton::Left,
+            2 => MouseButton::Middle,
+            3 => MouseButton::Right,
+            // buttons 4 through 7 are for scrolling.
+            8 => MouseButton::X1,
+            9 => MouseButton::X2,
+            _ => {
+                log::warn!("unknown mouse button code {}", button);
+                MouseButton::None
+            }
+        }
+    }
+
+    fn mouse_buttons(&self, mods: u16) -> MouseButtons {
+        let mut buttons = MouseButtons::new();
+        let button_masks = &[
+            (xcb::xproto::BUTTON_MASK_1, MouseButton::Left),
+            (xcb::xproto::BUTTON_MASK_2, MouseButton::Middle),
+            (xcb::xproto::BUTTON_MASK_3, MouseButton::Right),
+            (xcb::xproto::BUTTON_MASK_4, MouseButton::X1),
+            (xcb::xproto::BUTTON_MASK_5, MouseButton::X2),
+        ];
+        for (mask, button) in button_masks {
+            if mods & (*mask as u16) != 0 {
+                buttons.insert(*button);
+            }
+        }
+        buttons
+    }
+
+    fn key_mods(&self, mods: u16) -> KeyModifiers {
+        let mut ret = KeyModifiers::default();
+        let mut key_masks = [
+            (xcb::xproto::MOD_MASK_SHIFT, &mut ret.shift),
+            (xcb::xproto::MOD_MASK_CONTROL, &mut ret.ctrl),
+            // X11's mod keys are configurable, but this seems
+            // like a reasonable default for US keyboards, at least,
+            // where the "windows" key seems to be MOD_MASK_4.
+            (xcb::xproto::MOD_MASK_1, &mut ret.alt),
+            (xcb::xproto::MOD_MASK_4, &mut ret.meta),
+        ];
+        for (mask, key) in &mut key_masks {
+            if mods & (*mask as u16) != 0 {
+                **key = true;
+            }
+        }
+        ret
+    }
+
     pub fn handle_key_press(&self, key_press: &xcb::KeyPressEvent) -> Result<(), Error> {
         let key: u32 = key_press.detail() as u32;
         let key_code: KeyCode = key.into();
         let key_event = KeyEvent::new(
             key_code,
+            // TODO: detect repeated keys
             false,
-            KeyModifiers {
-                /// Shift.
-                shift: false,
-                /// Option on macOS.
-                alt: false,
-                /// Control.
-                ctrl: false,
-                /// Meta / Windows / Command
-                meta: false,
-            },
+            self.key_mods(key_press.state()),
             key_code,
             key_code,
         );
@@ -554,19 +597,17 @@ impl Window {
     }
 
     pub fn handle_button_press(&self, button_press: &xcb::ButtonPressEvent) -> Result<(), Error> {
+        let button = self.mouse_button(button_press.detail());
         let mouse_event = MouseEvent {
             pos: Point::new(button_press.event_x() as f64, button_press.event_y() as f64),
-            // TODO: Fill with held down buttons
-            buttons: MouseButtons::new().with(MouseButton::Left),
-            mods: KeyModifiers {
-                shift: false,
-                alt: false,
-                ctrl: false,
-                meta: false,
-            },
+            // The xcb state field doesn't include the newly pressed button, but
+            // druid wants it to be included.
+            buttons: self.mouse_buttons(button_press.state()).with(button),
+            mods: self.key_mods(button_press.state()),
+            // TODO: detect the count
             count: 1,
             focus: false,
-            button: MouseButton::Left,
+            button,
             wheel_delta: Vec2::ZERO,
         };
         match self.handler.try_borrow_mut() {
@@ -585,22 +626,19 @@ impl Window {
         &self,
         button_release: &xcb::ButtonReleaseEvent,
     ) -> Result<(), Error> {
+        let button = self.mouse_button(button_release.detail());
         let mouse_event = MouseEvent {
             pos: Point::new(
                 button_release.event_x() as f64,
                 button_release.event_y() as f64,
             ),
-            // TODO: Fill with held down buttons
-            buttons: MouseButtons::new(),
-            mods: KeyModifiers {
-                shift: false,
-                alt: false,
-                ctrl: false,
-                meta: false,
-            },
+            // The xcb state includes the newly released button, but druid
+            // doesn't want it.
+            buttons: self.mouse_buttons(button_release.state()).without(button),
+            mods: self.key_mods(button_release.state()),
             count: 0,
             focus: false,
-            button: MouseButton::Left,
+            button,
             wheel_delta: Vec2::ZERO,
         };
         match self.handler.try_borrow_mut() {
@@ -624,14 +662,8 @@ impl Window {
                 motion_notify.event_x() as f64,
                 motion_notify.event_y() as f64,
             ),
-            // TODO: Fill with held down buttons
-            buttons: MouseButtons::new(),
-            mods: KeyModifiers {
-                shift: false,
-                alt: false,
-                ctrl: false,
-                meta: false,
-            },
+            buttons: self.mouse_buttons(motion_notify.state()),
+            mods: self.key_mods(motion_notify.state()),
             count: 0,
             focus: false,
             button: MouseButton::None,