+++ CHANGELOG.md
@@ -77,6 +77,7 @@ While some features like the clipboard, menus or file dialogs are not yet availa
 - X11: Support individual window closing. ([#900] by [@xStrom])
 - X11: Support `Application::quit`. ([#900] by [@xStrom])
 - GTK: Support file filters in open/save dialogs. ([#903] by [@jneem])
+- X11: Support key and mouse button state. ([#920] by [@jneem])
 
 ### Visual
 
@@ -161,6 +162,7 @@ While some features like the clipboard, menus or file dialogs are not yet availa
 [#903]: https://github.com/xi-editor/druid/pull/903
 [#909]: https://github.com/xi-editor/druid/pull/909
 [#917]: https://github.com/xi-editor/druid/pull/917
+[#920]: https://github.com/xi-editor/druid/pull/920
 
 ## [0.5.0] - 2020-04-01